+++ Epsilon.h
@@ -10,9 +10,9 @@
 
 #include <iostream>
 
-inline int cmp(double x, double y) {
-  //static double epsilon = pow (10, - 10 * log10(x * y));
-  static double epsilon = 1e-10;
+inline int cmp(float x, float y) {
+  //static float epsilon = pow (10, - 10 * log10(x * y));
+  static float epsilon = 1e-10;
   if (x > y + epsilon) {
     return 1;
   }
