+++ MatrixTransformation.h
@@ -14,14 +14,14 @@ struct data_full_task {
   Matrix *matrix;
   int piv_col;
   int piv_row;
-  double *dev_col;
+  float *dev_col;
   Matrix *pin_matrix;
   d_matrix dev_matrix;
   cudaStream_t stream;
 
 };
 struct data_async {
-  double *dev_col;
+  float *dev_col;
   Matrix *pin_matrix;
   d_matrix dev_matrix;
   cudaStream_t stream;
@@ -37,9 +37,9 @@ struct data_async {
 };
 int matrixTransformCpu(Matrix &matrix, const int row, const int col);
 int matrixTransformAsync(Matrix &matrix, const int row, const int col, const int threads, data_async &data0, data_async &data1);
-int matrixTransformSync(Matrix &matrix, const int row, const int col, d_matrix &dev_matrix, double *dev_col);
+int matrixTransformSync(Matrix &matrix, const int row, const int col, d_matrix &dev_matrix, float *dev_col);
 
-__global__ void matrixTransform(d_matrix matrix, int piv_row, double *col_e);
+__global__ void matrixTransform(d_matrix matrix, int piv_row, float *col_e);
 int matrixTransformDouble (int threads, data_full_task data0, data_full_task data1);
 
 #endif /* MATRIXTRANSFORMATION_H_ */
