+++ TransitionMatrix.h
@@ -12,7 +12,7 @@
 
 int dev_trans_init(d_matrix &dev_trans, Matrix &input);
 int dev_trans_free(d_matrix &dev_trans);
-__global__ void fill_right_trans(d_matrix matrix, int col, double *row);
+__global__ void fill_right_trans(d_matrix matrix, int col, float *row);
 void modifyTransMatrAsync (int flag, int pivot_row, int pivot_col, d_matrix &temp_trans_1, d_matrix &temp_trans_2,
     d_matrix right_temp, cudaStream_t str_tr_ma);
